+++ Runtime/_Core/StringExtensions.cs
@@ -12,9 +12,7 @@
         /// <returns>An integer representing the FNV-1a hash of the input string.</returns>
         public static int ComputeFNV1aHash(this string str) {
             uint hash = 2166136261;
-            foreach (char c in str) {
-                hash = (hash ^ c) * 16777619;
-            }
+            foreach (char c in str) hash = (hash ^ c) * 16777619;
             return unchecked((int)hash);
         }
     }